switch to float32

+++ generativeMLP.py
@@ -75,7 +75,7 @@ def buildNet(input_var, layerNum, inDim, hidden, outDim, useReLU):
     return l_out
 
 def sampleInitial(n, inDim):
-    return np.random.normal(loc=0.0, scale=1.0/4, size=(n, inDim))
+    return np.random.normal(loc=0.0, scale=1.0/4, size=(n, inDim)).astype(np.float32)
     # The old cubemixture model is now unreachable:
     discrete = np.random.randint(0, 2, (n, inDim))
     continuous = np.random.normal(loc=0.0, scale=1.0/4, size=(n, inDim))
